requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.7
pydantic-settings
python-dotenv==1.0.0
Pillow==10.1.0
ultralytics==8.1.0
opencv-python==4.8.1.78